Bug 1: I was getting this message: TypeError: Fetch Failed Solution: I had disconnected the laptop from the internet. Reconnecting the Wi-fi resolved the issue.
Bug 2: I was seeing a Invalud Input data when inputting a recipe for upload
Solution: I had missed the call parentheses on the trim method inside the fix for ensuring my data has only space in between Quant/Unit/Description

Bug 3: I was finding that after moving my clearBookmarks function into a file of its own, with the intention of importing and calling it, as and when I needed it, but keeping it out of my main code, that I received an error message saying it was not a function.

This is not correct, as the function works when used inline inside of the model.js file.

The exporting of the function appears to be doing something to it, and changing how it is handled by the app.

Solution: For now, I have decided to leave it in model.js as the instructor has done and may revisit this later on..