BUG: Recipe Success Message Appeared after opening modal a second time. FIX: Add in location.reload() to the end of controller function for this feature. This reloaded the controlAddRecipe function, and meant that a new form was created whenever this button was clicked. The Success message then, get overridden, and the form shows up instead.
BUG: The quantity units were too close to the values, making the string of an ingredient more difficult to read.
FIX: I located the quantity class in the CSS and reduced it in size.

IMP: I would like to take this second BUG further and better target the units I wish to be flush with the value and the ones I did not. If a recipe calls for 2KG of Flour, then it should be margin-right: 0 but if a Recipe calls for 1 Large Egg, there should be a space.

This will be looked at a later date.

For now though, I am happy the project is finished. I will make one last change to the Placeholder text in the form, to finish it off.